完成作业

+++ Assignment1/src/rasterizer.cpp
@@ -2,16 +2,14 @@
 // Created by goksu on 4/6/19.
 //
 
-#include <math.h>
+#include "assignment/rasterizer.hpp"
 
 #include <algorithm>
+#include <math.h>
 #include <opencv2/opencv.hpp>
 #include <stdexcept>
 
-#include "assignment/rasterizer.hpp"
-
-rst::pos_buf_id rst::rasterizer::load_positions(
-    const std::vector<Eigen::Vector3f>& positions)
+rst::pos_buf_id rst::rasterizer::load_positions(const std::vector<Eigen::Vector3f> &positions)
 {
     auto id = get_next_id();
     pos_buf.emplace(id, positions);
@@ -19,8 +17,7 @@ rst::pos_buf_id rst::rasterizer::load_positions(
     return {id};
 }
 
-rst::ind_buf_id rst::rasterizer::load_indices(
-    const std::vector<Eigen::Vector3i>& indices)
+rst::ind_buf_id rst::rasterizer::load_indices(const std::vector<Eigen::Vector3i> &indices)
 {
     auto id = get_next_id();
     ind_buf.emplace(id, indices);
@@ -130,40 +127,38 @@ void rst::rasterizer::draw_line(Eigen::Vector3f begin, Eigen::Vector3f end)
     }
 }
 
-auto to_vec4(const Eigen::Vector3f& v3, float w = 1.0f)
+auto to_vec4(const Eigen::Vector3f &v3, float w = 1.0f)
 {
     return Vector4f(v3.x(), v3.y(), v3.z(), w);
 }
 
-void rst::rasterizer::draw(rst::pos_buf_id pos_buffer,
-                           rst::ind_buf_id ind_buffer, rst::Primitive type)
+void rst::rasterizer::draw(rst::pos_buf_id pos_buffer, rst::ind_buf_id ind_buffer, rst::Primitive type)
 {
     if (type != rst::Primitive::Triangle)
     {
-        throw std::runtime_error(
-            "Drawing primitives other than triangle is not implemented yet!");
+        throw std::runtime_error("Drawing primitives other than triangle is not implemented yet!");
     }
-    auto& buf = pos_buf[pos_buffer.pos_id];
-    auto& ind = ind_buf[ind_buffer.ind_id];
+    auto &buf = pos_buf[pos_buffer.pos_id];
+    auto &ind = ind_buf[ind_buffer.ind_id];
 
     float f1 = (100 - 0.1) / 2.0;
     float f2 = (100 + 0.1) / 2.0;
 
     Eigen::Matrix4f mvp = projection * view * model;
-    for (auto& i : ind)
+    for (auto &i : ind)
     {
         Triangle t;
 
-        Eigen::Vector4f v[] = {mvp * to_vec4(buf[i[0]], 1.0f),
-                               mvp * to_vec4(buf[i[1]], 1.0f),
-                               mvp * to_vec4(buf[i[2]], 1.0f)};
+        Eigen::Vector4f v[] = {
+            mvp * to_vec4(buf[i[0]], 1.0f), mvp * to_vec4(buf[i[1]], 1.0f), mvp * to_vec4(buf[i[2]], 1.0f)
+        };
 
-        for (auto& vec : v)
+        for (auto &vec : v)
         {
             vec /= vec.w();
         }
 
-        for (auto& vert : v)
+        for (auto &vert : v)
         {
             vert.x() = 0.5 * width * (vert.x() + 1.0);
             vert.y() = 0.5 * height * (vert.y() + 1.0);
@@ -185,24 +180,24 @@ void rst::rasterizer::draw(rst::pos_buf_id pos_buffer,
     }
 }
 
-void rst::rasterizer::rasterize_wireframe(const Triangle& t)
+void rst::rasterizer::rasterize_wireframe(const Triangle &t)
 {
     draw_line(t.c(), t.a());
     draw_line(t.c(), t.b());
     draw_line(t.b(), t.a());
 }
 
-void rst::rasterizer::set_model(const Eigen::Matrix4f& m)
+void rst::rasterizer::set_model(const Eigen::Matrix4f &m)
 {
     model = m;
 }
 
-void rst::rasterizer::set_view(const Eigen::Matrix4f& v)
+void rst::rasterizer::set_view(const Eigen::Matrix4f &v)
 {
     view = v;
 }
 
-void rst::rasterizer::set_projection(const Eigen::Matrix4f& p)
+void rst::rasterizer::set_projection(const Eigen::Matrix4f &p)
 {
     projection = p;
 }
@@ -215,15 +210,14 @@ void rst::rasterizer::clear(rst::Buffers buff)
     }
     if ((buff & rst::Buffers::Depth) == rst::Buffers::Depth)
     {
-        std::fill(depth_buf.begin(), depth_buf.end(),
-                  std::numeric_limits<float>::infinity());
+        std::fill(depth_buf.begin(), depth_buf.end(), std::numeric_limits<float>::infinity());
     }
 }
 
 rst::rasterizer::rasterizer(int w, int h) : width(w), height(h)
 {
-    frame_buf.resize(w * h);
-    depth_buf.resize(w * h);
+    frame_buf.resize(static_cast<long>(w) * h);
+    depth_buf.resize(static_cast<long>(w) * h);
 }
 
 int rst::rasterizer::get_index(int x, int y)
@@ -231,13 +225,10 @@ int rst::rasterizer::get_index(int x, int y)
     return (height - y) * width + x;
 }
 
-void rst::rasterizer::set_pixel(const Eigen::Vector3f& point,
-                                const Eigen::Vector3f& color)
+void rst::rasterizer::set_pixel(const Eigen::Vector3f &point, const Eigen::Vector3f &color)
 {
     // old index: auto ind = point.y() + point.x() * width;
-    if (point.x() < 0 || point.x() >= width || point.y() < 0 ||
-        point.y() >= height)
-        return;
+    if (point.x() < 0 || point.x() >= width || point.y() < 0 || point.y() >= height) return;
     auto ind = (height - point.y()) * width + point.x();
     frame_buf[ind] = color;
 }
